+++ src/de/omegazirkel/risingworld/adminutils/exports/PluginExport.java
@@ -1,8 +1,0 @@
-package de.omegazirkel.risingworld.adminutils.exports;
-
-public record PluginExport(
-        int id,
-        String name,
-        String path,
-        int loadOrder) {
-}